Use common initialization
