+++ core/include/scipp/core/element/creation.h
@@ -6,6 +6,7 @@
 
 #include <limits>
 
+#include "scipp/common/initialization.h"
 #include "scipp/common/overloaded.h"
 #include "scipp/core/element/arg_list.h"
 #include "scipp/core/subbin_sizes.h"
@@ -24,10 +25,8 @@ constexpr auto zeros_not_bool_like =
                  using T = std::decay_t<decltype(x)>;
                  if constexpr (std::is_same_v<T, bool>)
                    return int64_t{0};
-                 else if constexpr (std::is_same_v<Eigen::Vector3d, T>)
-                   return T::Zero();
                  else
-                   return T{0};
+                   return zero_init<T>::value();
                }};
 
 template <class T, T Value>
